start moving cloudinit/net functions into distro networking classes
